Fixed bug with fd reset followed by destruction after close

+++ toolbelt/fd.cc
@@ -7,7 +7,7 @@ void CloseAllFds(std::function<bool(int)> predicate) {
   struct rlimit lim;
   int e = getrlimit(RLIMIT_NOFILE, &lim);
   if (e == 0) {
-    for (int fd = 0; fd < lim.rlim_cur; fd++) {
+    for (int fd = 0; fd < static_cast<int>(lim.rlim_cur); fd++) {
       if (fcntl(fd, F_GETFD) == 0 && predicate(fd) ) {
         (void)close(fd);
       }
